+++ src/OpenTelemetry.Instrumentation.Process/CHANGELOG.md
@@ -5,7 +5,8 @@
 * Updated semantic conventions to
   [v1.42.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.42.0/docs/system/process-metrics.md).
   ([#4602](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4602))
-  * The `process.cpu.time` metric attribute `process.cpu.state` was renamed to `cpu.mode`.
+  * **Breaking Change**: The `process.cpu.time` metric attribute `process.cpu.state`
+    was renamed to `cpu.mode`.
   * Added the `process.uptime` metric.
   * Added the `process.windows.handle.count` metric (Windows only).
   * Added the `process.unix.file_descriptor.count` metric (non-Windows only).